requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor
pydantic==2.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0